[#5145] Move aggregation jobs history modal
